Cleanup

+++ src/main/java/party/lemons/biomemakeover/entity/render/BlightBatRender.java
@@ -1,8 +1,6 @@
 package party.lemons.biomemakeover.entity.render;
 
-import net.minecraft.client.render.entity.BatEntityRenderer;
 import net.minecraft.client.render.entity.EntityRenderDispatcher;
-import net.minecraft.client.render.entity.EntityRenderer;
 import net.minecraft.client.render.entity.MobEntityRenderer;
 import net.minecraft.client.render.entity.model.BatEntityModel;
 import net.minecraft.client.render.entity.model.EntityModel;
@@ -12,7 +10,6 @@ import net.minecraft.util.Identifier;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.MathHelper;
 import party.lemons.biomemakeover.BiomeMakeover;
-import party.lemons.biomemakeover.entity.BlightbatEntity;
 import party.lemons.biomemakeover.entity.render.feature.BlightbatFeatureRenderer;
 
 public class BlightBatRender extends MobEntityRenderer<BatEntity, EntityModel<BatEntity>> {
@@ -40,7 +37,7 @@ public class BlightBatRender extends MobEntityRenderer<BatEntity, EntityModel<Ba
         if (batEntity.isRoosting()) {
             matrixStack.translate(0.0D, -0.10000000149011612D, 0.0D);
         } else {
-            matrixStack.translate(0.0D, (double)(MathHelper.cos(f * 0.3F) * 0.1F), 0.0D);
+            matrixStack.translate(0.0D, MathHelper.cos(f * 0.3F) * 0.1F, 0.0D);
         }
 
         super.setupTransforms(batEntity, matrixStack, f, g, h);
